Add footer with links to pages such as privacy
